+++ Story/TripStore.swift
@@ -94,8 +94,8 @@ extension Day {
         if  let date = dict["date"] as? NSDate,
             let image = dict["image"] as? String,
             let text = dict["text"] as? String,
-            let latitude = dict["latitude"] as? Float64,
-            let longitude = dict["longitude"] as? Float64 {
+            let latitude = dict["latitude"] as? Float,
+            let longitude = dict["longitude"] as? Float {
             return Day(text: text, image: Image(name: image, date: date, latitude: latitude, longitude: longitude))
         }
         return nil
